fix(sedimentdrift): times_resuspended must not be an 8-bit counter
Found while writing the validation notebooks, when the same quantity came out as
44 in one analysis and 254 in another. SedimentElement.times_resuspended was
declared np.uint8, so it wraps at 255: the live element array showed a mean of
44.6 lift-offs for the coarse class in the Sherwood double-resuspension case,
while the exported history peaked at exactly 255.0. Neither number was real. With
the counter widened to uint32 the true mean is 298.8, with a maximum of 341.

The other uint8 element variables are genuine 0/1 flags or small class indices
and are left alone; this was the only unbounded counter stored in 8 bits.

This reaches past the test suite: times_resuspended is in the export_variables
list of the DDT production runs, so resuspension counts in existing output are
wrapped wherever a particle exceeded 255 lift-offs -- which, given the hopping
defect the campaign found, the fast classes certainly did.

+++ opendrift/models/sedimentdrift.py
@@ -140,7 +140,10 @@ class SedimentElement(Lagrangian3DArray):
                       'units': '1',
                       'default': 1}),
         # keeps track of how many times a particle was resuspended
-        ('times_resuspended', {'dtype': np.uint8,
+        # An unbounded counter, so it must not be uint8: a coarse grain on an
+        # energetic bed can be lifted hundreds of times in a few days, and an
+        # 8-bit counter silently wraps at 255, corrupting the diagnostic.
+        ('times_resuspended', {'dtype': np.uint32,
                       'units': '1',
                       'default': 0}),
         ('beached', {'dtype': np.uint8,  # 0 is active, 1 is settled
